+++ src/main/java/com/tripmoa/TripMoaApplication.java
@@ -1,13 +1,13 @@
-package com.tripmoa.backend;
+package com.tripmoa;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BackendApplication {
+public class TripMoaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
+		SpringApplication.run(TripMoaApplication.class, args);
 	}
 
 }
