--- src/main/java/com/tripmoa/backend/test/ConnectTestController.java
@@ -1,14 +1,0 @@
-package com.tripmoa.backend.test;
-
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ConnectTestController {
-
-    @GetMapping("/connect-test")
-    public String connectTest() {
-        return "backend connected";
-    }
-}
