+++ src/main/java/com/tripmoa/test/controller/testController.java
@@ -1,0 +1,4 @@
+package com.tripmoa.test.controller;
+
+public class testController {
+}
