+++ src/main/java/com/tripmoa/test/domain/test.java
@@ -1,0 +1,4 @@
+package com.tripmoa.test.domain;
+
+public class test {
+}
