+++ src/test/java/com/tripmoa/TripMoaApplicationTests.java
@@ -1,10 +1,10 @@
-package com.tripmoa.backend;
+package com.tripmoa;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BackendApplicationTests {
+class TripMoaApplicationTests {
 
 	@Test
 	void contextLoads() {
